requirements/main.txt
-i https://pypi.org/simple
django-dotenv==1.4.2
django-filter==2.1.0
django==2.1.8
djangorestframework==3.9.4
markdown==3.1.1
psycopg2-binary==2.8.2
pytz==2019.1
django-cors-headers==2.2.0
djangorestframework-jwt==1.11.0
django-vanilla-views==1.0.6
python-dateutil==2.8.0
django-taggit==1.1.0
django-model-utils==3.2.0